+++ Changelog.md
@@ -1,5 +1,16 @@
 # Release History
 
 
+
+### 1.1.0 ( 2024-03-06)
+- [x] Library works with v.3.0.1 of `@peter.naydenov/shortcuts` package
+- [x] New method `enablePlugin` to enable the shortcut plugins;
+- [x] New method `disablePlugin` to disable the shortcut plugins;
+- [x] Package.json: "exports" section was added. Allows you to use package as commonjs or es6 module without additional configuration;
+- [x] Folder `dist` was added to the package;
+- [x] Rollup was added to the project. Used to build the library versions;
+
+
+
 ## 0.0.1 ( 2023-11-23 )
 - [x] Initial release